requirements.txt
discord.py
flask
tzdata